+++ R/abr1.R
@@ -47,10 +47,12 @@
 #' dim(dat)
 #' 
 #' # names of the variables
-#' dimnames(dat)[[2]]
+#' dimnames(dat)[[2]] %>% 
+#'  head()
 #' 
 #' # print out the experimental factors
-#' print(abr1$fact)
+#' abr1$fact %>% 
+#'  head()
 #' 
 #' # check out the repartition of class
 #' table(abr1$fact$class)