+++ R/internals.R
@@ -2,7 +2,7 @@
 remote_repository <- 'aberHRML/metaboData'
 
 #' @importFrom piggyback pb_list
-#' @importFrom fs fs_bytes
+#' @importFrom fs fs_bytes dir_create
 
 remoteData <- function(remote_repository){
     pb_list(repo = remote_repository) %>%
@@ -32,9 +32,26 @@ dataSetAvailable <- function(technique,dataSet,dataSetDir,internalDir){
     
     if (!(technique %in% available_data_sets$technique) | 
         !(dataSet %in% available_data_sets$`data set`)){
-       FALSE
+        stop('Data set not available. Use "availableDataSets()" to find available data sets.',call. = FALSE)
     } else {
-        TRUE
+        invisible(TRUE)
+    }
+}
+
+filesAvailable <- function(files,technique,dataSet,dataSetDir,internalDir){
+    available_files <- availableFiles(technique,
+                                          dataSet,
+                                          dataSetDir = dataSetDir,
+                                          internalDir = internalDir)
+    
+    if (FALSE %in% (files %in% available_files$file_name)){
+        
+        incorrect_files <- files[!(files %in% available_files$file_name)] %>% 
+            str_c(collapse = ', ')
+        
+        stop(str_c('Files ',incorrect_files,' are not available. Use "availableFiles()" to find available files for this data set.'),call. = FALSE)
+    } else {
+        invisible(TRUE)
     }
 }